+++ CHANGELOG.md
@@ -4,7 +4,10 @@
     - Add orderVariablesForResolution method for ordering type variable declarations
     - findVariables now finds preserved variables (ExplicitTypeVariable) too  
 * Improve types tracking: tracked types now analyzed for dependent variables to extract all possible type information
-* Fix reversed generic variables declaration support (#3)       
+* Fix reversed generic variables declaration support (#3)
+* TypeResolutionUtils
+    - Add shortcut resolve(Class, LinkedHashMap<String, Type>, Class...) for resolution with known root generics
+    - Add resolve(Type, Class...) to support resolution from ParameterizedType (and to be used as universal resolution method)      
 
 ### 3.0.0 (2018-06-19)
 * Add constructor generics support